backend_dj/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
pillow==10.4.0
python-dotenv==1.0.1
python-telegram-bot==21.5
sniffio==1.3.1
sqlparse==0.5.1
